feat: final stake history update
