+++ apps/frontend/src/app/3_organisms/StakingHistoryFrame/StakingHistoryFrame.tsx
@@ -1,10 +1,10 @@
 import React, { FC, useCallback, useState } from 'react';
 
-import { StakingDelegateChanges } from './StakingDelegateChanges';
-import { StakingExtendedDuration } from './StakingExtendedDuration';
-import { StakingHistory } from './StakingHistory/StakingHistory';
 import { StakingistoryType } from './StakingHistoryFrame.type';
-import { StakingWithdrawns } from './StakingWithdrawns';
+import { StakingDelegateChanges } from './components/StakingDelegateChanges';
+import { StakingExtendedDuration } from './components/StakingExtendedDuration';
+import { StakingHistory } from './components/StakingHistory/StakingHistory';
+import { StakingWithdrawns } from './components/StakingWithdrawns';
 
 export const StakingHistoryFrame: FC = () => {
   const [selectedHistoryType, setSelectedHistoryType] = useState(
