+++ apps/frontend/src/app/3_organisms/StakingHistoryFrame/components/StakingHistory/StakingHistory.constants.tsx
@@ -3,27 +3,25 @@ import React from 'react';
 import { t } from 'i18next';
 
 import { SupportedTokens } from '@sovryn/contracts';
-import { Paragraph, ParagraphSize } from '@sovryn/ui';
 
-import { AmountRenderer } from '../../../2_molecules/AmountRenderer/AmountRenderer';
-import { TransactionIdRenderer } from '../../../2_molecules/TransactionIdRenderer/TransactionIdRenderer';
-import { getTokenDisplayName } from '../../../../constants/tokens';
-import { translations } from '../../../../locales/i18n';
-import { dateFormat } from '../../../../utils/helpers';
+import { AmountRenderer } from '../../../../2_molecules/AmountRenderer/AmountRenderer';
+import { TransactionIdRenderer } from '../../../../2_molecules/TransactionIdRenderer/TransactionIdRenderer';
+import { getTokenDisplayName } from '../../../../../constants/tokens';
+import { translations } from '../../../../../locales/i18n';
+import { dateFormat } from '../../../../../utils/helpers';
 import { V2StakingHistoryItem } from './StakingHistory.types';
 
-export const generateRowTitle = (item: V2StakingHistoryItem) => (
-  <Paragraph size={ParagraphSize.small} className="text-left">
-    {dateFormat(item.timestamp)}
-  </Paragraph>
-);
-
-export const columnsConfig = [
+export const COLUMNS_CONFIG = [
   {
     id: 'timestamp',
     title: t(translations.common.tables.columnTitles.timestamp),
     cellRenderer: (tx: V2StakingHistoryItem) => dateFormat(tx.timestamp),
     sortable: true,
+  },
+  {
+    id: 'transactionType',
+    title: t(translations.common.tables.columnTitles.transactionType),
+    cellRenderer: () => t(translations.stakingHistory.increase),
   },
   {
     id: 'lockedUntil',
@@ -33,10 +31,11 @@ export const columnsConfig = [
   },
   {
     id: 'amount',
-    title: t(translations.stakingHistory.amount),
+    title: t(translations.stakingHistory.stakeChange),
     cellRenderer: (tx: V2StakingHistoryItem) => (
       <AmountRenderer
         value={tx.amount || 0}
+        prefix="+"
         suffix={getTokenDisplayName(SupportedTokens.sov)}
       />
     ),
