+++ apps/frontend/src/app/3_organisms/StakingHistoryFrame/components/StakingHistory/index.ts
