+++ apps/frontend/src/app/3_organisms/StakingHistoryFrame/components/StakingWithdrawns/StakingWithdrawns.utils.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { t } from 'i18next';
+
+import { Paragraph, ParagraphSize } from '@sovryn/ui';
+
+import { translations } from '../../../../../locales/i18n';
+import { dateFormat } from '../../../../../utils/helpers';
+import { V2StakingWithdrawnItem } from './StakingWithdrawns.types';
+
+export const generateRowTitle = (item: V2StakingWithdrawnItem) => (
+  <Paragraph size={ParagraphSize.small} className="text-left">
+    {t(translations.stakingHistory.unstake)}
+    {' - '}
+    {dateFormat(item.timestamp)}
+  </Paragraph>
+);
